+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## WIP
+
+### Added
+- Added node access check
+
 ## 1.0.0 (April 13, 2016)
 
 ## 0.10.0 (December 15, 2015)